Separate Webview into its own subdir, move widgets into a subdirectory.
