+++ lib/scarpe.rb
@@ -9,31 +9,12 @@ require "securerandom"
 require "json"
 
 require_relative "scarpe/version"
-
 require_relative "scarpe/display_service"
+require_relative "scarpe/wv/widget" # REMOVE THIS LINE LATER
+require_relative "scarpe/widgets"
 
-require_relative "scarpe/widget"
-require_relative "scarpe/app"
-require_relative "scarpe/colors"
-require_relative "scarpe/dimensions"
-require_relative "scarpe/html"
-
-require_relative "scarpe/spacing"
-require_relative "scarpe/para"
-require_relative "scarpe/background"
-require_relative "scarpe/border"
-require_relative "scarpe/stack"
-require_relative "scarpe/flow"
-require_relative "scarpe/button"
-require_relative "scarpe/image"
-require_relative "scarpe/edit_box"
-require_relative "scarpe/edit_line"
-require_relative "scarpe/alert"
-
-require_relative "scarpe/text_widget"
-require_relative "scarpe/link"
-
-require_relative "scarpe/document_root"
+# WebView Display Service
+require_relative "scarpe/wv"
 
 class Scarpe
   class << self
