+++ lib/scarpe/widgets.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+require_relative "widget"
+require_relative "app"
+
+require_relative "spacing"
+require_relative "background"
+require_relative "border"
+
+require_relative "document_root"
+require_relative "para"
+require_relative "stack"
+require_relative "flow"
+require_relative "button"
+require_relative "image"
+require_relative "edit_box"
+require_relative "edit_line"
+require_relative "alert"
+
+require_relative "text_widget"
+require_relative "link"
