+++ lib/scarpe/wv.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+# Scarpe Webview Display Service
+
+require_relative "wv/widget"
+
+require_relative "wv/colors"
+require_relative "wv/dimensions"
+require_relative "wv/html"
+
+require_relative "wv/spacing"
+require_relative "wv/background"
+require_relative "wv/border"
+
+require_relative "wv/app"
+require_relative "wv/document_root"
+require_relative "wv/para"
+require_relative "wv/stack"
+require_relative "wv/flow"
+require_relative "wv/button"
+require_relative "wv/image"
+require_relative "wv/edit_box"
+require_relative "wv/edit_line"
+require_relative "wv/alert"
+
+require_relative "wv/text_widget"
+require_relative "wv/link"
