+++ lib/scarpe/wv/colors.rb
