+++ lib/scarpe/wv/dimensions.rb
