+++ lib/scarpe/wv/html.rb
